SCC-41 syntax analyzer decomposition
